started working on the buttons for advancing and reducing list position for images
